+++ src/main/java/com/wellnesswave/WellnessWave/Repository/DiagnosisRepository.java
@@ -3,5 +3,5 @@ package com.wellnesswave.WellnessWave.Repository;
 import com.wellnesswave.WellnessWave.Entities.Diagnosis;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DiagnosisRepository extends JpaRepository<Diagnosis, Long> {
+public interface DiagnosisRepository extends JpaRepository<Diagnosis, Integer> {
 }
